Better random number generator for velocity and position initialization

+++ clib/random.h
@@ -2,5 +2,5 @@
 
 #include <stdlib.h>
 
-void c_random_seed(unsigned int seed);
+void c_random_seed(long int seed);
 double c_random();
